STYLE VIEW LAYER
Orginally, the application displayed the title and url with no styling. I added in styling utilizing Bootstrap to make the information that is being displayed more readable. I adding in some color, I put the information into a card to make it more modularized. Lastly, I gave styling to the button to make the overall layout and appearance more appealing to the reader's eye.